Fixed add-product bug. Fixed the controller structure.

+++ sap/src/main/java/com/project/sap/controllers/ProductsController.java
@@ -1,48 +1,75 @@
 package com.project.sap.controllers;
 
-import com.project.sap.models.Laptop;
+import com.project.sap.models.*;
 import com.project.sap.models.Dto.LaptopDto;
-import com.project.sap.models.Sale;
-import com.project.sap.services.AddComponentsService;
-import com.project.sap.services.LaptopService;
-import com.project.sap.services.SalesService;
+import com.project.sap.services.interfaces.AddComponentsService;
+import com.project.sap.services.interfaces.LaptopService;
+import com.project.sap.services.interfaces.SalesService;
+import com.project.sap.utils.InputValidation;
+import com.project.sap.utils.LaptopMapperResolver;
+import org.apache.commons.lang3.math.NumberUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
-//TODO Make pages, so that not all info is loaded at once
 @Controller
 public class ProductsController {
 
-    @Autowired
+    private SalesService salesService;
+    private LaptopService laptopService;
     AddComponentsService addComponentsService;
+    LaptopMapperResolver laptopMapper;
+
     @Autowired
-    LaptopService laptopService;
-    @Autowired
-    SalesService salesService;
-    @GetMapping("/products")
-    public ModelAndView getProducts(){
-        ModelAndView mav = new ModelAndView("products");
-
-        List<Laptop> laptopList = laptopService.get();
-
-        mav.addObject("products", laptopList);
-        return mav;
-
+    public ProductsController(LaptopService laptopService, AddComponentsService addComponentsService, LaptopMapperResolver laptopMapper, SalesService salesService) {
+        this.laptopService = laptopService;
+        this.addComponentsService = addComponentsService;
+        this.laptopMapper = laptopMapper;
+        this.salesService = salesService;
     }
 
-    //TODO make deleting products not possible when there is a sale on the certain product or possible but give the user a warning beforehand
+    @GetMapping("/add-product")
+    public ModelAndView addProduct(Model model){
+        ModelAndView mav = new ModelAndView("addproduct");
+        mav.addObject("ramsList", addComponentsService.getRams());
+        mav.addObject("processorList", addComponentsService.getProcessors());
+        mav.addObject("screenList", addComponentsService.getScreens());
+        mav.addObject("storageList", addComponentsService.getStorages());
+        mav.addObject("videoCardList", addComponentsService.getVideoCards());
+        mav.addObject("laptop", new LaptopDto());
+        return mav;
+    }
+
+    @PostMapping("/add-product")
+    public ModelAndView saveProduct(LaptopDto laptopDto, Model model){
+        InputValidation.lenghtCheck("Manufacturer",laptopDto.getManufacturer(), model);
+        InputValidation.lenghtCheck("Model", laptopDto.getModel(), model);
+        InputValidation.numberCheck("Price", laptopDto.getPrice(), model);
+        if (model.getAttribute("error") != null){
+            return this.addProduct(model);
+        }
+        Laptop laptop = laptopMapper.laptopDtoToLaptop(laptopDto);
+        laptopService.add(laptop);
+        return new ModelAndView("redirect:/home");
+    }
+
     @GetMapping("/delete/products/{id}")
-    public String deleteUser(@PathVariable(value = "id") String id){
+    public String deleteUser(@PathVariable(value = "id") String id, RedirectAttributes ra){
         long currentId = Long.parseLong(id);
-        laptopService.deleteById(currentId);
+        if (salesService.saleWithProductExists(currentId)){
+            ra.addFlashAttribute("error", "You cannot delete a laptop who has a sale!");
+        }else {
+            laptopService.deleteById(currentId);
+        }
         return "redirect:/products";
     }
 
@@ -51,9 +78,7 @@ public class ProductsController {
         long currentId = Long.parseLong(id);
         Laptop laptop = laptopService.findById(currentId).get();
         ModelAndView mav = new ModelAndView("edit-product");
-        LaptopDto laptopDto = laptopService.mapLaptopToDto(laptop);
-        //String ramIds = "[" + laptopDto.getRamIds().stream().map(String::valueOf).collect(Collectors.joining(",")) + "]";
-        //String storageIds = "[" + laptopDto.getStorageIds().stream().map(String::valueOf).collect(Collectors.joining(",")) + "]";
+        LaptopDto laptopDto = laptopMapper.laptopToLaptopDto(laptop);
         mav.addObject("imageList", laptop.getImages());
         mav.addObject("product", laptopDto);
         mav.addObject("ramsList", addComponentsService.getRams());
@@ -66,13 +91,31 @@ public class ProductsController {
     }
 
     @PostMapping("edit/products/{id}")
-    public String submitEditProduct(@PathVariable(value = "id") String id, LaptopDto laptopDto){
-        Laptop laptop = laptopService.mapDtoToLaptop(laptopDto);
+    public String submitEditProduct(@PathVariable(value = "id") String id, LaptopDto laptopDto, RedirectAttributes ra){
+        if (!NumberUtils.isCreatable(laptopDto.getPrice()) || Double.parseDouble(laptopDto.getPrice())<=0){
+            ra.addFlashAttribute("error", "The price cannot be equal or less than 0!");
+            return "redirect:/edit/products/" + id;
+        }
+        Laptop laptop = laptopMapper.laptopDtoToLaptop(laptopDto);
+        laptop.setSales(laptopService.findById(laptopDto.getId()).get().getSales());
         laptop.setId(laptopDto.getId());
         for (Sale sale: salesService.getAll().stream().filter(x->x.getLaptop().getId()==laptop.getId()).collect(Collectors.toList())) {
             sale.setTotalPrice(BigDecimal.valueOf(laptop.getPrice()*sale.getQuantity()));
+            salesService.save(sale);
         }
+
         laptopService.add(laptop);
         return "redirect:/products";
+    }
+
+    @GetMapping("/products")
+    public ModelAndView getProducts(){
+        ModelAndView mav = new ModelAndView("products");
+
+        List<Laptop> laptopList = laptopService.get();
+
+        mav.addObject("products", laptopList);
+        return mav;
+
     }
 }
